removes use of Named Pipe
I swapped out the problematic named pipe for a socket variant. 

The socket version was a bit tricky because I wanted to make sure it was secure. I did this 3 fold.

The first thing I did was i made it run on the internal loopback of the TCP/IP stack. this will ensure that only something on the local computer can access it. 

The second thing I did was the data sent across the connection is encrypted using AES.

The Third thing was a bit tricky. What I have set up is the generation of the key used is not something anyone is going to have easy access to. and the key gets changed each and every single build. the key is inserted into the CryptoSocket file and that file is then compiled into a dll using Cython. All traces of the source file containing the key are deleted immediately. and the compiled dll is then added to the compile lib folder.

here is an issue with Stackles and Cython. I have worked this problem out and have a specially compiled version of the Stackless Python27.dll file. This file is included in the build/data/python27 folder.
Because cythonizing the CryptoSocket code causes an error if the builder is run using the stock Stackless dll I have the builder do some checks to see if the dll is located in the python installations folder. if it is the file gets renamed. and then the path to the new dll is added to the environment. and the build gts restarted as a subprocess. not before setting a marker in the environment so we do not end up in a loop. after the build is finished. successful or not the original dll gets put back into place. If the dll is not found in the installation folder. then the pah to the new one is added and away we go.