removed running the build process in a thread when the GUI is not used.
I removed this for 2 reasons. Number one is exceptions don't always
cross thread apartments if there is an error the build process simply
terminates with no log data. the second reason is that there are wx
modules that get used in the build process. and wx get somewhat picky
about using it's modules for anything other then the main thread. This
causes EG to hang t the end of the build because of an error that wx
fires off about using one of it's modules outside of the man thread.
This error doe not come in the form of an exception. But instead it
comes as a message box the user has to click on. Kinda hard to close
it on appveyor.