default language revisited
Thi is the second time doing this. the first ime I could not remember
what the problem was exactly. It is something that is not easily
replicated.

Details:

the description of wx.Locale.GetSystemLanguage states
"tries to detect users default locale setting."
This description is inaccurate. It does not try to get the "users"
default locale. it does exactly what the method name implies.
it gets the "system" default language. which may not be the language
the currently logged in user has set to be displayed o their screen.

The system default language is determined by what was purchased. So if
i live in Russia. the Windows available for purchase there would be a
russian edition. Now as the user I cna tell windows to display whatever
language I want. so to me everything is pretty seamless. If we
installed EG on a Russian copy of Windows where the user has all of the
language settings set to English, EG wuold display the Russian language.
Not Ideal. We want to display the language that the user is
currently displaying in Windows.