Removes Cython and AES Encryption
There appears to have been some issues with the build process
using Cython. So I have removed Cython. I made an effort to build my
own AES module I am going to leave the module in with this PR as I
would like to see If i can get it fixed.

The issue I am having with it I really do not know where the problem is.
If i set up script to bombard the module with 100+ things to encrypt or
decrypt it works without issue. But if i use the module from different
processes sometimes the data gets corrupted. Not sure where the issue is.
It is with allocation of memory I am sure of it.

So for the time being there is no AES encryption the socket does run on
the loopback so any attempt to hyjack anything would have to be done
from the local comeputer.

I did want to mention that i did pound this thing with 1000 processes
hitting it all at the same time. every single event came through. in
actually an astounding amount of time. 18 seconds. or 18 ms per event.