Fixes InnoSetup template to work with newer versions of InnoSetup
In newer versions of InnoSetup they dropped support for 2 languages,
These languages have been removed from the template.

There was also a line in the template that was Inc(i) It appears as
tho that command may have been in older versions of InnoSetup Unicode
In the current documentation it is not listed. So I am going to assume
it has been removed. It is the same thing as doing i := i + 1 so
it was not a big deal to change it.