Removes ctypeslib as a dependency.
From what I can find ctypeslib is only used to create WinApi.Dynamic
It is no longer being maintained. In order for it to run it requires.
GCC-XML which is also no longer being maintained. It has been succeeded
by Cast-XML which is not compatible with ctypes lib. you can download
an old version of GCC-XML the problem is that it only works with
MSVC 6, 7 or 7.1. There is a fork on Pypi that uses Clang, but it is
not complete and does not function for what EG would use it for.

what ctypeslib does is it creates python versions of C/C++ code.
What EG used it for was converting some of the Windows API.

I already have 500K lines of code of the Windows API converted from
Windows Vista to the latest build of Windows 10. So it is pretty much
obsolete and no longer needed.