OK so here is the ho down low down on this.
I am leveraging the Windows API directly to acquire all kinds of
goodies pertaining to locales. I still have to rely on the locale
module to set the locale properly. this iis because setlocale is a
compiler directive and not a Windows API function call. I would have
to write a python extension to call setlocale. which is what the
locale module already does.

Now here is the crafty bit. we really do not care what the windows
user locale is set to. we do however care what language EG is set to.

I created a series of country and language classes. These classes make
up all of the country/language combinations that windows supports from
XP to Windows 10 (latest release) including all of the language packs.

That is
254 different countries
270 different languages
with a grand total of 610 country/language combinations.

pretty cool eh?

on my windows 7 box with no language packs, there are
123 countries that Windows supports
123 languages that Windows supports
with a grand total of 196 country/language combinations that Windows supports

This is the part that is the boot in the backside. We have to remember
that wxPython is also involved here. and wxPython only supports
172 of the 610 Windows locals. and my Windows 7 box supports
127 of the 172.

because of the plethora of varying locale adjustments that are available
in Windows the use of locale's that do not have a Windows LCID attached
to iit is becoming more and more common. This has become an issue
because of the number of locale that are like this that are in Windows 10.

So how I have managed to skirt around this problem is this. I use the
Windows API to locate the code page for a language. then i also use the
Windows API to collect the english name of the language and the country.

I then format the 3 items like so

    english_language_name_english_country_name.codepage

and this is what I pass to `locale.setlocale`

because of the direct use of a codepage Windows does not rely on an LCID
to set the application locale.

So what I do is I use the iso code in eg.config match up the proper locale
for it. and then call set locale using the actual set language in EG.

I do feel that wee need to expand upon this and provide some additional
options. like setting the language and locale on their own. because the
reality is they really have not a thing to do with each other. the
"locale" or country component iis more for things like
numeric, calendar, currency, time and date formatting an the language
is simply the displayed text.

It should be fairly simple to add this ability. I have also added a bunch
of properties that return the different locale components. like
currency symbol, percent symbol.... so on and so forth. it gets these
directly from Windows for the language/locale that is currently set.