Fixes locale problems
There are multiple locale issues between wxPython and the locale module in Python. This is an attempt to correct 3 of those issues. The first one is with wxPython and GetSystemLocale not returning the correct locale. The second is with wxPythons use of LCID's to identify a locale. While I have not correct the use of the LCID's I have correct it's use of LCID's when using wx.Locale.GetInfo().

The 3rd problem is in Pythons locale module. The locale module does not create a proper locale string for setting the locale. If you pass "en_US" to locale.setlocale it will fail. with the changes I have made it will no longer fail. It also returns the correct locale string for windows. When using python 2.7 on windows it seems that you must use english language name_english country name.code page. So now if you pass en_US to locale.setlocale it is going to query windows and get the english names for the language and country and append a modified version of the code page that has been supplied if one was supplied.