adds better locale support
I updated the Options dialog so it will display all languages and locales that have translations coded into EG. I also expanded the number of flags to a total of 249.

The language selection now displays the language and country name in the native script if available, otherwise it is displayed in english.