adds better handling of the tolerances.
I made the tolerance operate in a more variable way. If a protocol has a lead in and a start bit I find out the percentage difference between the 2 and if that percentage is < 50 I then add in the user set tolerance. If the user set tolerance and the calculated one is < 40 i use that. if not I add the calculated tolerance and the user tolerance / 2. I only use this calculated tolerance for checking the data part of the code and also the lead out. the lead in is checked using the tighter user supplied tolerance.