hopefully fixes for detection of receiver, protocol decoding, tolerance, repeat code timing
I also improved upon the repeat code speed and duplicate code speed. I now have it down to 500 millionths of a second (1/2 a millisecond!!). There is no way to improve upon the new code protocol detection which is currently running at about 9 milliseconds. This is because it is a process of hunt and peck in order to find the proper decoder. That is really the only way that it can be done.

The MCE receiver timings are a HUGE pain because of how sloppy they are allowed to be. This is going to make it impossible to decode some of the protocols without disabling all of the other first. I believe I can make this process easier for the user if they use the database I am going to set up. If the code's friendly name does not match then they know that the remote is not going to be that protocol and thus can disable it. I am also going to attempt to institute a priority setting for the protocols once I manage to collect more information.

I did also add 2 new features. The user is now able to get the repeat count of a code and also the number of times a code has been used.

I created a new problem that I have to sort out. This deals with the saved named of a code.