Several Changes
Corrected a starting issue with the plugin. I forgot to remove a file write from the ioctrl.py file.
Cleans up the code_wrapper.CodeWrapper.__init__ method.
Fixes the name for an IRCode when decoded by the MCE decoder
Changes the order of the decoders.
Adds some additional fine tuned tolerances.
fixes the naming of some of the properties in the ir_code.IRCode class
corrects issue with xml_handler and text fields when the field is multi line.